backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
sentence-transformers>=2.7.0
dateparser>=1.2.0
PyYAML>=6.0.1
# IBM Docling for robust PDF/DOC parsing
# If installation fails on your platform, keep the fallback (pdfminer.six)
docling>=2.0.0
pdfminer.six>=20231228
python-docx>=1.0.0
